Refactored from comments on pull request. Also pulled the input comment box into it's own class
