+++ src/main/js/posts/post.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Comment from "./comments/comment";
-import Like from "./comments/like";
+import Like from "./likes/like";
 
 const Post = (props) => {
 	return (
